src/requirements.txt
click
colored
requests
python-socketio
